fix: remove shellbar content class
